Day 13 - Claw Contraption
Somehow, possibly because I was working on the train, I way overcomplicated this problem. Unlike the general disappointment on the internet I knew immediately that this was a systems of equations problem. Where I failed is thinking that the intersection of the two legs had to occur on a whole number point on the graph. Then I did some bad math and made things worse.

Once I stopeed and drew a picture I went back and ripped out all the useless solutions it ran on the first attempt.